Implement custom comment brightness customization
> Description

Added a accessibility theme customization to allow users to set the
brightness of comments with percentage adjustments from 0% - 20% where 0
is the default nord3 comment color. When a invalid value is
passed (< 0 || > 20 || nil) the default nord3 color is used.

The customization is available through the nord-comment-brightness variable.

This is a optional customization to provide a way for users to adjust
the color to fit their needs: It is a port of the reference
implementation from the Nord Atom UI project. The values are calculated
using the LESSCSS lighten function to ensure full interoperability with
other port projects that providing this theme feature.

> Usage

Example to increase the comment color brightness by 15%:
  (setq nord-comment-brightness 15)

References:
https://github.com/arcticicestudio/nord-atom-syntax/issues/47
http://lesscss.org/functions/#color-operations-lighten

GH-43