Implement region highlight style customization
> Description

Added a theme customization to allow users to set a style for the region
highlighting either to "snowstorm" or "frost", providing a way for users
to adjust the colors to fit their needs.

The default is left to "nord2" as background- and no specific foreground
color.

The customization is available through the "nord-region-highlight"
variable of type "string".

> Usage

Example using the "snowstorm" style:
  (setq nord-region-highlight "snowstorm")

Resolves #32

GH-45