Prefix internal names to avoid global namespace pollution
Sadly, Emacs Lisp has no namespaces, so every top level definition is
visible to the entire universe.  That includes variables and
functions.

This commit prefixes them to avoid "leaking" them in a way that could
conflict with other code.  The convention of `package--` is used to
mirror the most current Emacs coding conventions.

There are no functional changes from this modification.

Closes #66