Migrate away from deprecated `show-paren-X-face` faces
The `show-paren-match-face` and `show-paren-mismatch-face` face are
deprecated since Emacs version 22.1 [1,2] and were removed in Emacs 25.
To provide compatibilty to later EMacs version, the new
`show-paren-match` and `show-paren-mismatch` faces have been added.

The deprecated faces will be removed later on, but for now its totally
fine for both faces to co-exist to support Emacs versions <22.1 as well
as >=22.1.

[1]: http://git.savannah.gnu.org/cgit/emacs.git/commit/?id=c430f7e23fc2c22f251ace4254e37dea1452dfc3
[2]: https://github.com/emacs-mirror/emacs/commit/c430f7e23fc2c22f251ace4254e37dea1452dfc3

GH-75