Fix `Invalid face box` error (#130)
The `style` attribute `sunken-button` for the `:box` attribute set on a
`button` or `custom-button*` is not a valid value[1].

This was not made manifest by a bug[2] on the emacs side but is expected
to be fixed and become an error in the version 30 release.

In the context of the code, `sunken-button` itself is considered
equivalent to `pressed-button`, so all places where `sunken-button` is
specified have been replaced with `pressed-button`.

[1]: https://www.gnu.org/software/emacs/manual/html_node/elisp/Face-Attributes.html
[2]: https://debbugs.gnu.org/cgi/bugreport.cgi?bug=67404