Fix invalid face for `:box` error as of Emacs version 30
The style attribute `sunken-button` for the `:box` attribute set on a
`button` or `custom-button*` is not a valid value [1] anymore. It has
always been invalid due to a bug [2] in a previous version as is thrown
as an error as of Emacs version 30.

In the context of the code, `sunken-button` itself is considered
equivalent to `pressed-button`, so all places where `sunken-button` was
specified has been replaced with `pressed-button`.

[1]: https://www.gnu.org/software/emacs/manual/html_node/elisp/Face-Attributes.html
[2]: https://debbugs.gnu.org/cgi/bugreport.cgi?bug=67404

GH-131
Closes GH-130

Signed-off-by: Sven Greb <development@svengreb.de>